requirements.txt
# HubSpot CRM Bulk Import - Requirements
# Python 3.8+

# HubSpot API
requests==2.31.0

# Data processing
pandas==2.1.3

# Data validation
email-validator==2.1.0
phonenumbers==8.13.26

# Configuration
python-dotenv==1.0.0

# Reporting & Visualization
plotly==5.18.0
jinja2==3.1.2

# Progress tracking
tqdm==4.66.1

# Date handling
python-dateutil==2.8.2
